requirements.txt
future==0.18.0
hydra-core>=1.1
loguru>=0.3
numpy>=1.17
omegaconf>=2.1
scikit-learn>=1.0
tqdm>=4.50.0
jupyterlab==3.5.1
albumentations==1.3.1
open_clip_torch==2.23.0
einops==0.7.0
git+https://github.com/facebookresearch/segment-anything.git
transformers==4.29.2
diffusers==0.16.1
accelerate==0.20.1
xformers
pytorch-lightning==1.4.2
torchmetrics==0.6.0
kornia==0.6
matplotlib
mat73==0.62
pre-commit
hydra-submitit-launcher
pandas
seaborn
